ReRank/requirements.txt
Flask
cohere
weaviate-client
python-dotenv
